Move functions to seprate files
